requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.11
PyPDF2==3.0.1
sentence-transformers==2.6.1
torch
sqlitedict==2.1.0
google-generativeai==0.3.2
python-dotenv==1.0.1
